+++ example/ioc/setupServices.ts
@@ -6,8 +6,8 @@ import { FooApi } from '../api/FooApi';
 import { ExampleService } from '../service/ExampleService';
 
 export function setupServices() {
-  ServiceContainer.set(ServiceId.BarApi, () => new BarApi());
-  ServiceContainer.set(ServiceId.FooApi, () => new FooApi());
-  ServiceContainer.set(ServiceId.CustomApi, () => new CustomApi());
-  ServiceContainer.set(ServiceId.Service, () => new ExampleService());
+  ServiceContainer.set(ServiceId.BarApi, BarApi);
+  ServiceContainer.set(ServiceId.FooApi, FooApi);
+  ServiceContainer.set(ServiceId.CustomApi, CustomApi);
+  ServiceContainer.set(ServiceId.Service, ExampleService);
 }