+++ src/Service.ts
@@ -1,5 +1,0 @@
-export type Service = {
-  id: string;
-  factory: () => any;
-  instance?: any;
-}