+++ Changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2021-07.29
+
+* You are now able to initialize a class only by its reference
+  e.g. `ServiceContainer.set(ServiceId.TestApi, CustomTestApi);` instead
+  of `ServiceContainer.set(ServiceId.TestApi, () => new CustomTestApi();`
+
 ## [1.4.0] - 2021-05-31
 
 ### Added